Creates an API endpoint for GET all links
I built this endpoint out when I got to the stage, in the test, where the function
getLinks needs to make an API call. Instead of trying to work with the HTML,
I just made an API endpoint, using the existing links controller. Currently
stuck on trying to get a debugger to show me what data is being captured by the
GET request.