getLinks works in the console
getLinks test is not passing, which may be a database issue. However, it works,
in a rough way, in the console. Since the POST request, creating a link is
not yet working, it's not possible to create a link in the console yet. But
its a good start. the require 'jquery' was causing the biggest problem
for the getLinks method.