Completes filtering JS method
The method has a listener in the HTML for when a key is entered, and calls
the filterLinks method. That method iterates through the table rows and hides those that
do not include the letters in the input field. Test does not pass, even after
assets:precompile. Selenium issue continues.